refactor: extract onboarding logic from page.tsx into useOnboarding hook
Move all onboarding state, shared check injection, install gate, render
gates, and friend gate logic into a dedicated hook. page.tsx drops from
1506 to 1202 lines — just calls the hook and renders the result.
